hosts (raspberry-pi): remove bene user again
It turned out to be unpractical because the inital setup of a new
raspberry-pi uses the nixos user that's provisioned by the SD card
image. So some fundermental tools such as git, gh, and the initial
checkout of this repository are done by that user. Moving everything
over to the bene user just for the sake of user name consitency
is really not worth it.