hosts (latitude-7280): Fix incorrect swap device reference
I noticed this because the boot process was always hanging for 1m 30s
waiting for the device with UUID 277064c8-d24c-4948-a55e-fc6f6567544b to
appear. Looking at the result of `ll /dev/disk/by-uuid` I could see that
now device with that UUID exists in this system. Using `lsblk` I was
able to inspect the devices present in my system including sda2 which
has a size of 8GB - the size I initially configured for swap.

I suspect that this error slipped in when I reinstalled this system.
Implementing #15 seems to be a way of avoiding this in the future.