hosts (latitude-7280): Add disko setup
This prepares for using disko for formatting disks. This switches
from ext4 to btrfs in order to more easily adopt impermanence some
time in the future (see #16).

This disko configuration is at this point not imported anywhere.
I first need to boot using a live medium and then format the disk
using this configuration. Only then I can use this as my disk
configuration file.

Part of #15