home (alacritty): Pin theme file
Previously the code would fetch the catppuccin-mocha.yml file
from the latest commit of the GitHub repository. In the
meantime the repository was changed to contain the thema as
a toml file instead. This made the rebuild fail when I
reinstalled on latitude-7280.

The configuration now pins the exact revision so the yml version
of the file is guaranteed to be present.