hosts: Fix disko setup for work-macbook and raspberry-pi
Even if the disko module is disabled, nix will still look for a key
called `disko` because it's referenced from modules/disko/default.nix.
This caused building the configuration of work-macbook and raspberry-pi
to fail.
To work around this raspberry-pi now imports the disko nixos module
although it does not enable disko. work-macbook does no longer
import all modules. This is probably the most pragmatic workaround
because most nixos modules are not going to work on darwin anyways.