home (java): Make configurable
Java is added to the list of default imported home modules.
In order to wire this up correctly for all hosts, the option
definitions for my.host and my.user had to be moved to a
common module called 'options'. This way it's safe to assume
they are present on all system.