home (git): Read user identifiy from osConfig
Adds email and signingKey entries to my.user in the os-level configuration.
These values are then used to configure the git home module. This removes the
need for a custom home-manager configuration for host raspberry-pi and allows
for dropping git/work-identity.nix. git/private-identity.nix is still required
due to the fact that it configures multiple directory specific identities, and
I still need to figure out how to model that.