home: Get rid of per host home entry points
Previously I stored host specific home-manager configuration in
dedicated entry point files. This resulted in the requirement to
have a dedicated home-manager nix module setup per host in that
host's configuration.nix.
After this change, hosts can declare profiles in addition to their
role. These roles enable profile-specific configurations in
home-manager. This allows for dropping the host specific home-manager
entry points.
For now there's just a work and a private profile, but there might
be more in the future. Since profiles are configured on the host
level and not on the home-manager level, they can also be utilized
to conditionally enable nixos modules on nixos hosts.

Resolves #10