hosts (cyberoffice): use sops to manage secrets
Secrets are stored in the secrets.yaml file alongside the host's configuration.
Keys permitted to read the secrets are configured in the top-level .sops.yaml
file.
When recreating the host, the corresponding key needs to be send during the
nix-anywhere installation process. To make that more straight forward, a
setup script is added in scripts folder.