modules (nginx): Introduce a fallback server for nginx
Setting up proxying between directions and srv-prod-2 only worked as
long as there was just a single virtual server defined on srv-prod-2.
When collabora was added as a secondary service, collabora would always
respond even when asking for nextcloud.ritter.family.
This is because the proxy_pass directive also sets the host header to
, but nginx on srv-prod-2 can't find a match for that. The
default behavior is to return whatever the first virtual server is,
that is defined in the nginx config.

By adding an explicit fallback server that always responds with 404,
this should be made more explicit and therefore easier to get to the
bottom of, should it happen again.

The result of deploying this change will be that nextcloud won't be
available via nextcloud.ritter.family anymore, even though collabora
is currently disabled. This is because the default host will serve
unmatched requests to nextcloud.ritter.family made to srv-prod-2.
In order to fix this, nextcloud.ritter.family has to be added as
a secondary server name for the vhost on srv-prod-2. This will
be in a follow up commit, once the new behavior was confirmed
to be working correctly.