Revert "modules (nginx): Introduce a fallback server for nginx"
Unfortunately this doesn't work for SSL connections because those need
an SSL certificate. One way of fixing this would be generating a
self-signed certificate, but for now I've just decided to fix the root
cause of the issue, which is srv-prod-2 needs to have vhost aliases for
the external DNS names.

This reverts commit bd0a7c45da51a2e5011bc26ef0e3bd5d1d33720f.