modules (https-proxy): Add aliases
This way proxied services can specify their external DNS name.
This is required because the external proxy will set the host header
to the host that was requested resulting in e.g. srv-prod-1 being
asked for host nextcloud.ritter.family.
Before this change this resulted in no match in src-prod-1's nginx
config which in turn fell back to the first server in the configuration.
This would then let that server respond to any requests being send
to srv-prod-1.