hosts (pulse-14): Restore defaulting for DHCP
Apparently although `networking.useDHCP` defaults to true in
hardware-configuration.nix, it's turned off again due to the fact that
`networking.networkmanager` is enabled in my custom networking module.
This commit restores the configuration setting that was previously
present in hardware-configuration.nix because setting the value to
true results in a conflict.