modules (vaultwarden): Add workaround for password.ritter.family
This is yet another workaround for the problem described in
bd0a7c45da51: When accessing password.ritter.family collabora is
currently responding because the host header from proxy forwarding from
directions is set to password.ritter.family and that does not match the
vhost defined for vaultwarden.
This change adds an alias so the vaultwarden vhost will serve requests.
I'm using the nextcloud.stage option as a workaround to make this
working for the moment. This also required defining a fallback for
stage. Without the fallback it would always complain that
my.modules.nextcloud.stage is used but not defined.