home (java): Correctly configure additional java installs for Gradle
It looks like `org.gradle.java.installations.paths` has been configured
incorrectly for quite a while. This is because `pkgs.jdkxx` is a
reference to the root of the java package, but that is not the Java
home. Instead Java home is inside `lib/openjdk` (at least for
`pkgs.jdkxx`). Luckily the jdk packages expose a home that can be used
to reference the Java home instead:
https://github.com/NixOS/nixpkgs/blob/9d9c9b1bfdf8700ac538f32177189d2496779ae2/pkgs/development/compilers/openjdk/generic.nix#L613

I't looks like more recent versions of Gradle started to look for the
tools.jar inside the lib folder and that resulted in Gradle build
crashing when updating the Gradle wrapper in some of my projects. The
fix is to use `pkgs.jdkxx.home` which is the "real" java home inside the
package.