hosts (srv-test-2): Fix nextcloud-sync script
The script was trying to connect via srv-prod-2's IP address but the ssh
config uses 'srv-prod-2' as host identifier. After this change, the
script uses the host identifier from the ssh configuration and also does
no longer pass the user name to connect with, because that's also
configured in ssh config.