home (nvim): Tweak tab behavior
The tabstop setting defined the width of a tab and defaults to 8 spaces.
This is too wide for my taste, so it's configured to 4 spaces instead.
softtabstop defines how many spaces are inserted when the tab key is
pressed. For the go programming language which uses tabs for
indentation, it is set to match the tabstop setting so a single tab
press will insert a tab character instead of first inserting two spaces
and then on second press replacing that with a tab.