flake: Update to nixpkgs 25.05
This includes setting `nixvim.useGlobalPackages` which was introduced in
nixvim 25.05 and behaves essentially like the setting of the same name
for Home Manager.

The point of this is to tell nixvim whether or not it should inherit and
reuse the nixpkgs instance passed down from Home Manager or not. While
the documentation states it is recommended to have nixvim manage it's
own nixpkgs instance, it resulted in very confusing errors during the
25.05 update: Rebuilding the system failed because terraform was not
part of the allowedUnfree packages, although it was (check
modules/home-manager/default.nix). This is because nixvim started to
bring in it's own instance of nixpkgs but the allowedUnfree setting was
only applied on instance defined for the NixOS system.

Because this really confused me, I decided to keep a single nixpkgs
instance and configure allowedUnfree packages that way.