hosts (pulse-14): Replace workaround for running protoc
For building TestLens using Gradle, protoc and the gRPC protoc plugin
are required. This used to work by linking ld-linux-x86-64.so.2 to
/lib64. After an upgrade of the protoc-grpc-java plugin, the build
started failing because it couldn't find libstdc++.so.6 anymore.

Instead of linking that as well, I discovered nix-ld, which fixes the
problem more idiomatically. See https://github.com/nix-community/nix-ld

Kudos to @marcphilipp for finding this.