flake: Base config on nixos-unstable
Instead of the latest stable release, the config is now based on nixos-unstable.
While this might require fixing deprecations more frequently, there won't be
a need for a big migration when a new nixos release happens.
Furthermore I get the most recent packages and don't have to wait half a year
in case package major releases can't be backported from unstable to stable.