hosts (srv-prod-2): Setup new restic backup
This uses minIO as the backup backend instead of sftp. Furthermore
It's going to be repository per service instead of repository per
server, which makes it more flexible when I want to split services
to new hosts.

This commit starts that idea with the most simple backup which is
the git repository.