hosts (srv-test-2): Fix git restore service
- align naming with backup services
- align configuration with backup service, including cache dir, unit
  dependencies, etc.
- Configure restore to root directory with an include for all files
  included in the snapshot.

The last item is particularly important. We want the restore to put the
files into their real location, hence we need to restore to root. We
can't restore to a different location because for large backups, such as
nextcloud it would require excessive file copying. We also can't use
`lastest:/path/to/file --target /path/to/file` because for vaultwarden
and nextcloud we have multiple paths in the backup. Running multiple
restore commands, one for each path would we an option, but there's the
danger of a snapshot being created in the middle resulting in files
being restored from different snapshots. This is because we don't lock
the repository because the backup user is read only.
The solution is to restore to the root directory, but due to the use of
`--delete` we also need to specify the paths we want to restore.
Otherwise restic fails with an error.