hosts (srv-prod-2, srv-test-2): align pg versions
Use pg_dump and pg_restore from the postgresql package used by the
postgresql service. Otherwise when refering to pkg.postgresql in
scripts, it will render the path to the default postgresql version
(currently 17) which is different from the version used by
services.postgresql (currently 15). This resulted in backups being
created with pg_dump 17 which failed on import with pg_restore 15.