modules (git-server): Use restrictive git-shell
This disallows interactive ssh sessions. The module was configured this
way initially (43a7101fccf36aad2361037e5013a82871bbb655) but it was
changed to enable the app-sync which required interactive ssh login.
Since the app-sync now works via minio and restic, it's not required
anymore to allow interactive ssh login for the git user.