modules (restic-restore): Run restore post command in ExecStartPost
If the post command is run in ExecStopPost, it counts towards TimeoutStopSec
which is 90s by default. Since restore post commands frequently do time
consuming operations suchs as restoring a database, 90s is not enough causing
the unit to be marked as failed.
Running the restore post command in ExecStartPost avoids this problem because
TimeoutStartSec has a default of infinity.
It's also more symetric to run the pre command in ExecStartPre and the post
command in ExecStartPost.